intial work for extenders scs
