+++ packages/zowe-explorer/i18n/sample/src/ZoweExplorerExtender.i18n.json
@@ -1,0 +1,3 @@
+{
+  "displayName": "Zowe Explorer"
+}
