+++ packages/zowe-explorer/i18n/sample/src/extension.i18n.json
@@ -8,9 +8,5 @@
   "activate.didSaveText.isDataSet": "File is a data set-- saving ",
   "activate.didSaveText.isUSSFile": "File is a USS file -- saving",
   "activate.didSaveText.file": "File ",
-  "activate.didSaveText.notDataSet": " is not a data set or USS file ",
-  "profile.init.read.imperative": "Unable to read imperative file. ",
-  "initialize.module.load": "Credentials not managed, unable to load security file: ",
-  "moveTempFolder.error": "Error encountered when creating temporary folder! ",
-  "deactivate.error": "Unable to delete temporary folder. "
+  "activate.didSaveText.notDataSet": " is not a data set or USS file "
 }
