+++ packages/zowe-explorer/i18n/sample/src/utils/CoreUtils.i18n.json
@@ -1,0 +1,6 @@
+{
+  "profile.init.read.imperative": "Unable to read imperative file. ",
+  "initialize.module.load": "Credentials not managed, unable to load security file: ",
+  "moveTempFolder.error": "Error encountered when creating temporary folder! ",
+  "deactivate.error": "Unable to delete temporary folder. "
+}
